cleanup extensions project to allow for publishing compiler packages
