+++ Source/GenerateSharp/Build.Extensions/IValueTableExtensions.cs
@@ -2,9 +2,6 @@
 // Copyright (c) Soup. All rights reserved.
 // </copyright>
 
-using System.Collections.Generic;
-using System.Linq;
-
 namespace Soup.Build
 {
 	public static class IValueTableExtensions
@@ -27,33 +24,6 @@ namespace Soup.Build
 			}
 
 			return table[propertyName].AsList();
-		}
-
-		public static void SetAll(this IValueList list, IEnumerable<IValue> values)
-		{
-			list.Clear();
-			foreach (var value in values)
-				list.Add(value);
-		}
-
-		public static void SetAll(this IValueList list, IValueFactory factory, IEnumerable<string> values)
-		{
-			list.Clear();
-			foreach (var value in values)
-				list.Add(factory.Create(value));
-		}
-
-		public static void Append(this IValueList list, IEnumerable<IValue> values)
-		{
-			foreach (var value in values)
-				list.Add(value);
-		}
-
-
-		public static void Append(this IValueList list, IValueFactory factory, IEnumerable<string> values)
-		{
-			foreach (var value in values)
-				list.Add(factory.Create(value));
 		}
 	}
 }
