+++ Source/GenerateSharp/Build.Extensions/Soup.Build.Extensions.csproj
@@ -6,5 +6,6 @@
 	</PropertyGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\Build\Soup.Build.csproj" />
+		<ProjectReference Include="..\Opal\Opal.csproj" />
 	</ItemGroup>
 </Project>
