+++ Source/GenerateSharp/Extensions/CSharp/Compiler/Core/BuildEngine.cs
@@ -3,7 +3,6 @@
 // </copyright>
 
 using Opal;
-using Soup.Build.Utilities;
 using System;
 
 namespace Soup.Build.CSharp.Compiler
