+++ Source/GenerateSharp/Extensions/CSharp/Compiler/Core/Recipe.toml
@@ -1,4 +1,4 @@
-Name = "Soup.Cpp.Compiler"
+Name = "Soup.CSharp.Compiler"
 Language = "C#"
 Version = "0.1.1"
 Source = [
@@ -7,7 +7,6 @@ Source = [
 	"BuildResult.cs",
 	"CompileArguments.cs",
 	"ICompiler.cs",
-	"LinkArguments.cs",
 	"MockCompiler.cs",
 ]
