+++ Source/GenerateSharp/Extensions/CSharp/Compiler/Roslyn/Compiler.cs
@@ -3,7 +3,6 @@
 // </copyright>
 
 using Opal;
-using Soup.Build.Utilities;
 using System.Collections.Generic;
 
 namespace Soup.Build.CSharp.Compiler.Roslyn
