+++ Source/GenerateSharp/Extensions/CSharp/Compiler/Roslyn/Recipe.toml
@@ -1,4 +1,4 @@
-Name = "Soup.Cpp.Compiler.MSVC"
+Name = "Soup.CSharp.Compiler.Roslyn"
 Language = "C#"
 Version = "0.1.1"
 Source = [
@@ -9,6 +9,6 @@ Source = [
 [Dependencies]
 Runtime = [
 	"Opal@1.0.1",
-	"Soup.Cpp.Compiler@0.1.1",
+	"Soup.CSharp.Compiler@0.1.1",
 	"Soup.Build.Extensions@0.1.5",
 ]
