+++ Source/GenerateSharp/Extensions/Cpp/Compiler/Core/BuildEngine.cs
@@ -3,7 +3,6 @@
 // </copyright>
 
 using Opal;
-using Soup.Build.Utilities;
 using System;
 using System.Collections.Generic;
