+++ Source/GenerateSharp/Extensions/Cpp/Compiler/MSVC/ArgumentBuilder.cs
@@ -3,7 +3,6 @@
 // </copyright>
 
 using Opal;
-using Soup.Build.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
