+++ Source/GenerateSharp/Extensions/Cpp/Compiler/MSVC/Compiler.cs
@@ -3,11 +3,9 @@
 // </copyright>
 
 using Opal;
-using Soup.Build.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace Soup.Build.Cpp.Compiler.MSVC
 {
